% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f7page.R
\name{f7_page}
\alias{f7_page}
\title{Mobile page wrapper}
\usage{
f7_page(..., navbar, toolbar, title = NULL, options = shinyMobile_options)
}
\arguments{
\item{...}{Content.}

\item{navbar}{Slot for \link{f7_navbar}.}

\item{toolbar}{Slot for \link{f7_toolbar}.}

\item{title}{Tab title.}

\item{options}{Options for Framework7}
}
\description{
Mobile page wrapper
}
